WIP: group_tree files renamed to flow_network
